[Chore] Rename home to auth
